requirements.txt
cmdstanpy==1.2.4
convertdate==2.4.0
dask==2024.10.0
holidays==0.59
importlib_resources==6.4.5
kaggle==1.6.17
lunarcalendar==0.0.9
matplotlib==3.9.2
plotly==5.24.1
protobuf==5.28.3
pytest==8.3.3
python_dateutil==2.8.2
setuptools==65.5.0
streamlit==1.39.0
tavily_python==0.5.0
google-ai-generativelanguage==0.6.10
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
tqdm==4.66.6
wheel==0.44.0
